:recycle: API polish
